refactor: Split Numeric core extensions out.
These are not used directly by Time::Iterator, but are for convenience.
They may be removed later.
