+++ lib/time/iterator/core_ext/numeric.rb
@@ -1,0 +1,7 @@
+# Inject 3.quarters
+class Numeric
+  def quarter
+    (3 * self).months
+  end
+  alias quarters quarter
+end
